requirements.txt
asgiref==3.8.1
attrs==25.3.0
begins==0.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.12.0
django-erd==1.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dot==0.3.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
eralchemy==1.5.0
funcsigs==1.0.2
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
pydot==3.0.2
PyJWT==2.10.0
pyparsing==3.2.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
setuptools==78.1.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uv==0.8.0
wheel==0.45.1